backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai-whisper==20231117
whisperx==3.1.1
pyannote.audio==3.1.1
torch>=2.1.2
torchaudio>=2.1.2
numpy>=1.24.3,<2.0
scipy>=1.11.4
librosa==0.10.1
pydantic==2.5.3
python-dotenv==1.0.0
soundfile==0.12.1
pydub==0.25.1
